修改蓝足球满分跳过

+++ exam/src/main/java/com/feipulai/exam/activity/basketball/BasketBallSetting.java
@@ -20,13 +20,14 @@ public class BasketBallSetting {
      * 满分跳过
      */
     private boolean fullSkip = false;
-    private int maleFullScore;//男子
-    private int femaleFullScore;//女子
+    private double maleFullScore;//男子
+    private double femaleFullScore;//女子
     /**
      * 分组测试模式 0 连续 1 循环
      */
-    private int testPattern =TestConfigs.GROUP_PATTERN_SUCCESIVE;
+    private int testPattern = TestConfigs.GROUP_PATTERN_SUCCESIVE;
     private int testType = 0;
+
     public int getTestNo() {
         return testNo;
     }
@@ -84,19 +85,19 @@ public class BasketBallSetting {
         this.fullSkip = fullSkip;
     }
 
-    public int getMaleFullScore() {
+    public double getMaleFullScore() {
         return maleFullScore;
     }
 
-    public void setMaleFullScore(int maleFullScore) {
+    public void setMaleFullScore(double maleFullScore) {
         this.maleFullScore = maleFullScore;
     }
 
-    public int getFemaleFullScore() {
+    public double getFemaleFullScore() {
         return femaleFullScore;
     }
 
-    public void setFemaleFullScore(int femaleFullScore) {
+    public void setFemaleFullScore(double femaleFullScore) {
         this.femaleFullScore = femaleFullScore;
     }
